refactor(ui): api notebook funcs
