+++ services/dashboard-ui/client/lib/ctl-api/notebooks/delete-cell.ts
@@ -1,0 +1,14 @@
+import { api } from '@/lib/api'
+import type { INotebookScoped } from './types'
+
+export const deleteCell = ({
+  orgId,
+  installId,
+  notebookId,
+  cellId,
+}: INotebookScoped & { cellId: string }) =>
+  api<void>({
+    orgId,
+    method: 'DELETE',
+    path: `installs/${installId}/notebooks/${notebookId}/cells/${cellId}`,
+  })
