+++ services/dashboard-ui/client/lib/ctl-api/notebooks/delete-notebook.ts
@@ -1,0 +1,13 @@
+import { api } from '@/lib/api'
+import type { INotebookScoped } from './types'
+
+export const deleteNotebook = ({
+  orgId,
+  installId,
+  notebookId,
+}: INotebookScoped) =>
+  api<void>({
+    orgId,
+    method: 'DELETE',
+    path: `installs/${installId}/notebooks/${notebookId}`,
+  })
