+++ services/dashboard-ui/client/lib/ctl-api/notebooks/get-notebook.ts
@@ -1,0 +1,8 @@
+import { api } from '@/lib/api'
+import type { TNotebook, INotebookScoped } from './types'
+
+export const getNotebook = ({ orgId, installId, notebookId }: INotebookScoped) =>
+  api<TNotebook>({
+    orgId,
+    path: `installs/${installId}/notebooks/${notebookId}`,
+  })
