+++ services/dashboard-ui/client/lib/ctl-api/notebooks/get-notebooks.ts
@@ -1,0 +1,8 @@
+import { api } from '@/lib/api'
+import type { TNotebook, IInstallScoped } from './types'
+
+export const getNotebooks = ({ orgId, installId }: IInstallScoped) =>
+  api<TNotebook[]>({
+    orgId,
+    path: `installs/${installId}/notebooks`,
+  })
