+++ services/dashboard-ui/client/lib/ctl-api/notebooks/update-cell.ts
@@ -1,0 +1,19 @@
+import { api } from '@/lib/api'
+import type { TNotebookCell, INotebookScoped } from './types'
+import type { ICreateCellBody } from './create-cell'
+
+export type IUpdateCellBody = Partial<ICreateCellBody>
+
+export const updateCell = ({
+  orgId,
+  installId,
+  notebookId,
+  cellId,
+  body,
+}: INotebookScoped & { cellId: string; body: IUpdateCellBody }) =>
+  api<TNotebookCell>({
+    orgId,
+    method: 'PATCH',
+    path: `installs/${installId}/notebooks/${notebookId}/cells/${cellId}`,
+    body,
+  })
